requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
argcomplete==3.2.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
crashtest==0.4.1
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.7
django-allauth==0.60.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-qr-code==4.0.0
djangorestframework==3.14.0
dulwich==0.21.7
fastjsonschema==2.19.0
filelock==3.13.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
Pillow==10.1.0
pipx==1.3.3
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
promise==2.3
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pyproject_hooks==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
rapidfuzz==3.6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
segno==1.6.0
service-identity==23.1.0
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomlkit==0.12.3
trove-classifiers==2023.11.29
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
userpath==1.9.1
virtualenv==20.25.0
wheel==0.42.0
zope.interface==6.1
